refactor: consolidate SUM/AVG/COUNT-DISTINCT into one scalar-agg scan
The three single-predicate row-scanning scalar aggregates each duplicated the
POST-leaflet scan scaffolding (resolve predicate, enumerate leaves, open handle,
skip empty/foreign leaflets). Fold them into one fast_predicate_scalar_agg
module: a shared scan_predicate_scalar_agg driver plus an AggState accumulator
enum (Sum / Avg / CountDistinct) that keeps each variant's per-leaflet gating,
per-row fold, and finalization. Output shapes are preserved exactly — SUM and
COUNT(DISTINCT) emit xsd:integer, AVG emits xsd:double (or Unbound on empty
input).

SUM's SumExprI64 expression machinery moves here from fast_fused_scan_sum (now
deleted); AVG leaves fast_min_max_string, which keeps the directory-only,
O(leaflets) MIN/MAX paths that would regress through a per-row cursor.
