+++ fluree-db-query/src/lib.rs
@@ -36,7 +36,6 @@ pub mod exists;
 pub mod explain;
 pub(crate) mod fast_count;
 pub(crate) mod fast_exists_join_count_distinct_object;
-pub(crate) mod fast_fused_scan_sum;
 pub(crate) mod fast_group_count_firsts;
 pub(crate) mod fast_label_regex_type;
 pub(crate) mod fast_min_max_string;
@@ -44,6 +43,7 @@ pub(crate) mod fast_multicolumn_join_count_all;
 pub(crate) mod fast_optional_chain_head_count_all;
 pub(crate) mod fast_path_common;
 pub(crate) mod fast_path_plus_count_all;
+pub(crate) mod fast_predicate_scalar_agg;
 pub(crate) mod fast_star_const_order_topk;
 pub(crate) mod fast_string_prefix_count_all;
 pub(crate) mod fast_sum_strlen_group_concat;
